Refactors package structure
