+++ app/src/main/java/com/picpay/desafio/android/data/repository/PicPayService.kt
@@ -1,5 +1,6 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.data.repository
 
+import com.picpay.desafio.android.data.entity.User
 import retrofit2.Call
 import retrofit2.http.GET
