+++ app/src/main/java/com/picpay/desafio/android/ui/contact/UserListAdapter.kt
@@ -1,9 +1,11 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.ui.contact
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.picpay.desafio.android.R
+import com.picpay.desafio.android.data.entity.User
 
 class UserListAdapter : RecyclerView.Adapter<UserListItemViewHolder>() {
